Move `FeedImageDataLoaderWithFallbackComposite` into production
